refactor(controller): move LottoController into lotto.controller package for clearer layering
