+++ src/main/kotlin/lotto/Application.kt
@@ -1,3 +1,6 @@
 package lotto
 
+import lotto.controller.LottoController
+
+
 fun main() = LottoController().run()
